pyproject.toml
[project]
name = "dairy-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "zakibadawi",email = "mog5808@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client (>=1.14.2,<2.0.0)",
    "openai (>=1.81.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "google-genai (>=1.16.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["migrations"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
